MTROPOLIS: Various fixes to get SPQR to boot
Among others: this adds SPQR (Windows) to the detection table, it adds a
stub for the ReturnModifier, adds some (currently unimplemented)
properties to the world manager, and fixes a bug in the handling of
relative paths in the ObjectReferenceVariableModifier.

This is absolutely not ready for production: more fixes are needed to
get SPQR to actually run past the second puzzle, and there is still
debug output everywhere.